requirements.txt
setuptools
black
pre-commit
gitlint
requests
jira[cli]
python-dotenv